saflora_ai_wrapper/pyproject.toml
[project]
name = "saflora-ai-wrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.2",
    "fastapi>=0.135.1",
    "groq>=0.37.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
    "uvicorn>=0.41.0",
]
